% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.adjmat}
\alias{as.adjmat}
\title{Create and assign a new class for use of \code{geomnet}'s \code{\link{fortify.adjmat}} function.}
\usage{
as.adjmat(dat)
}
\arguments{
\item{dat}{A square adjacency matrix.}
}
\description{
Create and assign a new class for use of \code{geomnet}'s \code{\link{fortify.adjmat}} function.
}

