% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bikes}
\alias{bikes}
\title{Bike sharing network (directed)}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of six variables of length 53:
\itemize{
  \item Start.station: Station where bike trip starts
  \item End.station: Station where bike trip ends
  \item n: Number of trips between the two stations
  \item minlength: ## Duration of shortest trip between the two stations (units?)
  \item min5perc: ## Fifth percentile of trip durations (units?)
  \item village: I'm not sure ###
}
\item the vertices data set consists of 16 variables with information on 21 stations:
\itemize{
  \item id: Station ID number
  \item name: Station name
  \item terminalName: ??? ###
  \item lastCommWithServer: ??? ###
  \item lat: Latitude of station location
  \item long: Longitude of station location
  \item installed: ??? ###
  \item locked: ??? ###
  \item installDate: ??? ###
  \item removalDate: ??? ###
  \item temporary: ??? ###
  \item public: ??? ###
  \item nbBikes: Number of bikes at the station
  \item nbEmptyDocks: Number of empty docks at the station
  \item latestUpdateTime: ??? ###
  \item village: ??? ###
}
}}
\usage{
bikes
}
\description{
A list of two datasets, vertices and edges, containing data on bike trips in the Washington DC metro area (see \url{https://secure.capitalbikeshare.com/}).
The variables are as follows:
}
\keyword{datasets}

