% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{email}
\alias{email}
\title{Email network (directed)}
\format{
A list of two data frames:
\itemize{
\item the edges data set consists of ten variables of length 9063:
\itemize{
  \item From: Email address of sender
  \item eID:  email ID. If an email ws sent to multiple recipients, the email ID is the same.
  \item Date: Date and time email was sent
  \item Subject: Subject line of email
  \item to: Email address of recipient
  \item month: Month email was sent
  \item day: Day of month email was sent
  \item year: Year email was sent
  \item nrecipients: Number of recipients of email
}
\item the nodes data set consists of 18 variables with information on 55 employees:
\itemize{
  \item label: Employee's email address
  \item LastName: Employee's last name
  \item FirstName: Employee's first name
  \item BirthDate: Employee's first date
  \item BirthCountry: Employee's (fictional) country of birth
  \item Gender: Employee's gender
  \item CitizenshipCountry: Employee's (fictional) country of citizenship
  \item CitizenshipBasis: Is citizenship by birth or parents?
  \item CitizenshipStartDate: When employee became a citizen
  \item PassportCountry: (Fictional) country issuing employee's passport
  \item PassportIssueDate: Date employee received passport
  \item PassportExpirationDate: Date employee's passport expires
  \item CurrentEmploymentType: Employee's department in the company
  \item CurrentEmploymentTitle: Employee's title at the company
  \item CurrentEmploymentStartDate: Date employee started at position
  \item MilitaryServiceBranch: Branch of the (fictional) military in which the employee serves
  \item MilitaryDischargeType: General or honorable discharge from military service?
}
}
}
\usage{
email
}
\description{
A list of two datasets, vertices and edges, containing data on employees and (fictitious)
email headers from two weeks of internal GAStech company email between employees made available as part of the VAST challenge 2014 (K. Cook, G. Grinstein, M. Whiting, see \url{http://hcil2.cs.umd.edu/newvarepository/benchmarks.php}).
The variables are as follows:
}
\keyword{datasets}
