% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{football}
\alias{football}
\title{College football games network (undirected)}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of three variables of length 613:
\itemize{
  \item
  \item
}
\item the vertices data set consists of two variables with information on 157 Iowa high schools:
\itemize{
  \item from: First school in game
  \item to: Second school in game
  \item same.conf: An indicator variable that is 1 if the two teams are in the same conference and 0 otherwise.
}
}}
\usage{
football
}
\description{
A list of two datasets, vertices and edges, containing data on Division I college football games in the Fall 2000 season.
The variables are as follows:
}
\references{
M. Girvan and M. E. J. Newman, Proc. Natl. Acad. Sci. USA 99, 7821-7826 (2002).
}
\keyword{datasets}

