% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-net.R
\name{fortify.adjmat}
\alias{fortify.adjmat}
\title{Function for converting a network adjacency matrix into the correct format for use with geomnet}
\usage{
\method{fortify}{adjmat}(model, data = NULL, ...)
}
\arguments{
\item{model}{An adjacency matrix of class \code{"adjmat"}.}

\item{data}{not used in this function}

\item{...}{not used in this function}
}
\description{
Function for converting a network adjacency matrix into the correct format for use with geomnet
}
\examples{
data(emon, package = "network")
adjmat <- as.adjmat(network::as.matrix.network.adjacency(emon$MtSi))
str(adjmat)
fortify(adjmat)
}

