% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-net.R
\name{fortify.matrix}
\alias{fortify.matrix}
\title{Function for converting a network adjacency matrix into the correct format for use with geomnet}
\usage{
\method{fortify}{matrix}(model, data = NULL, ...)
}
\arguments{
\item{model}{An adjacency matrix of class "matrix".}

\item{data}{NULL - not used in this function}

\item{...}{not used in this function}

\item{group}{character. Used for facetting. If you wish to facet on an edge variable provide the name of that variable here.}
}
\description{
Function for converting a network adjacency matrix into the correct format for use with geomnet
}
\examples{
data(emon, package = "network")
adjmat <- network::as.matrix.network.adjacency(emon$MtSi)
str(adjmat)
fortify(adjmat)
}

