% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-circle.r
\docType{data}
\name{GeomCircle}
\alias{GeomCircle}
\alias{geom_circle}
\title{Geom for drawing circles in the ggplot2 framework}
\format{\preformatted{Classes 'GeomCircle', 'Geom', 'ggproto' <ggproto object: Class GeomCircle, Geom>
    aesthetics: function
    default_aes: uneval
    draw_group: function
    draw_key: function
    draw_layer: function
    draw_panel: function
    non_missing_aes: 
    parameters: function
    required_aes: x y radius
    setup_data: function
    use_defaults: function
    super:  <ggproto object: Class Geom> 
}}
\usage{
GeomCircle

geom_circle(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Circles are drawn with a specified radius centered at (x, y).
This geom is very much exploratory - we are using it for drawing edges for self references.
It is not explored for any more general use, so use with caution!
}
\examples{
# circles are drawn centered at x and y
library(ggplot2)
data(mpg)
ggplot(mpg, aes(displ, hwy)) + geom_circle(size=0.1) + geom_point()
ggplot(mpg, aes(displ, hwy)) + geom_circle(linetype=2, size=0.05, alpha=0.5)
ggplot(mpg, aes(displ, hwy)) + geom_circle(aes(linetype=factor(cyl)), size=0.05, alpha=0.5)
}
\keyword{datasets}

