% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mm.directed}
\alias{mm.directed}
\title{A directed network of Mad Men relationships}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of two variables of length 87:
\itemize{
  \item Name1: Character variable with name of the character who made a sexual advance toward the character in Name2
  \item Name2: Character variable with name of the character receiving, not necessarily reciprocating, Name1's advance.
}
\item the vertices data set consists of two variables with information on 52 characters:
\itemize{
  \item label: Factor variable with name of the character
  \item Gender: Factor variable with gender of the character
}
}}
\usage{
mm.directed
}
\description{
A list of two datasets, vertices and edges, containing information on sexual advances made in the TV show Mad Men.
This data set was first compiled by Winston Chang for the package gcookbook (under the name \code{madmen2}) and was extended here to include
the gender of each of the characters.
The variables are as follows:
}
\references{
Winston Chang (2012). gcookbook: Data for "R Graphics Cookbook". R package
version 1.0. \url{http://CRAN.R-project.org/package=gcookbook}
}
\keyword{datasets}

