% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{protein}
\alias{protein}
\title{A protein interaction network (undirected)}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of two variables of length 4480:
\itemize{
  \item from: Protein 1 ID
  \item to: Protein 2 ID that interacts with protein 1
}
\item the vertices vector consists of the corresponding IDs of 2113 proteins in the edges data set
}}
\usage{
protein
}
\description{
A list of two datasets, vertices and edges, containing data on the complete protein-protein interaction network in the yeast species S. cerevisiae (see \url{http://www3.nd.edu/~networks/resources.htm}).
The variables are as follows:
}
\keyword{datasets}

