% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeNet.R
\name{reshapeNet}
\alias{reshapeNet}
\title{Function for converting various network structures into the correct format for use with geomnet}
\usage{
reshapeNet(net, ...)
}
\arguments{
\item{net.}{A network object. Can be a class "network", "igraph", "matrix" (for adjaceny matrices), or "data.frame" object. If net is a "data.frame", must provide the ndata argument and the first column of net must be the "from" node column.}

\item{ndata.}{Data frame containing network node list and other node information. Only applicable for class "data.frame." First column name should be node ids.}
}
\description{
Function for converting various network structures into the correct format for use with geomnet
}
\examples{

# class network (sna, network, statnet packages)

library(network)
data(emon)
reshapeNet(net = emon$Cheyenne)

# class igraph (igraph, igraphdata packages)
library(igraph)
library(igraphdata)
data("USairports", package = "igraphdata")
head(reshapeNet(net = USairports))

# class matrix (for adjacency matrices)

adjmat <- network::as.matrix.network.adjacency(emon$MtSi)
str(adjmat)
reshapeNet(net = adjmat)

# eclass data.frame and ndata 
data(blood)
reshapeNet(net = blood$edges, ndata = blood$vertices)

}

