% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{theme_elements}
\alias{theme_elements}
\title{\pkg{ggplot2} theme attribute inheritance network (directed)}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of two variables of length 48:
\itemize{
  \item child: Theme element that inherits its properties from the corresponding parent element
  \item parent: Theme element that passes its properties to its children
}
\item the vertices data set consists of one variable with information on 53 theme elements:
\itemize{
  \item name: Name of the theme element
}
}}
\usage{
theme_elements
}
\description{
A list of two datasets, vertices and edges, containing data on the inheritance structure of theme elements in \pkg{ggplot2} (see \url{http://docs.ggplot2.org/current/theme.html})
The variables are as follows:
}
\keyword{datasets}

