% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-net.r
\name{theme_net}
\alias{theme_net}
\title{Theme for network visualization}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Themes set the general aspect of the plot such as the colour of the
background, gridlines, the size and colour of fonts.
\code{theme_net} provides access to the regular ggplot2 theme, but removes any
background, axes, and ensures an aspect ratio of 1 for better
viewing of networks and graphs.
}
\examples{
library(ggplot2)
data(blood)
p <- ggplot(data = blood$edges, aes(from_id = from, to_id = to))
p + geom_net()
p + geom_net() + theme_net()

}

