% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.edgedf}
\alias{as.edgedf}
\title{Cast a data frame to an edgedf}
\usage{
as.edgedf(dat)
}
\arguments{
\item{dat}{A network edgelist of class \code{"data.frame"} for use with \code{\link{fortify.edgedf}}.}
}
\description{
Create and assign a new class for use of \code{geomnet}'s \code{\link{fortify.edgedf}} function.
}

