% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bikes}
\alias{bikes}
\title{Bike sharing network (directed)}
\format{A list of two data frames:
\itemize{
\item the trips data set consists of four variables of length 53:
\itemize{
  \item Start.station: Station where bike trip starts
  \item End.station: Station where bike trip ends
  \item n: Number of trips between the two stations
  \item minlength: Duration of shortest trip between the two stations (in seconds). Only those stations are included, if the shortest trip between them lasted not more than 15 minutes.
}

\item the vertices data set consists of five variables with information on 21 stations:
\itemize{
  \item id: Station ID number
  \item name: Station name
  \item lat: Latitude of station location
  \item long: Longitude of station location
  \item nbDocks: Number of bike docks at the station
}
}}
\usage{
bikes
}
\description{
This network is a summary of the bike trips taken by customers of the bike sharing company Capital Bikeshare (\url{https://secure.capitalbikeshare.com/}) during the second quarter of 2015.
Only trips between stations in the vicinity of Rockville, MD, are included.
The data is organized as a list of two datasets, vertices (stations) and edges (trips between stations),
as follows:
}
\references{
\url{https://secure.capitalbikeshare.com/}
}
\keyword{datasets}

