% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{football}
\alias{football}
\title{College football games network (undirected)}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of three variables of length 613:
\itemize{
  \item
  \item
}
\item the vertices data set consists of __ variables with information on __ actors:
\itemize{
  \item from: First school in game
  \item to: Second school in game
  \item same.conf: An indicator variable that is 1 if the two teams are in the same conference and 0 otherwise.
}
}}
\usage{
football
}
\description{
A list of two datasets, vertices and edges, containing data on Division I college football games played in Fall 2000 (see \url{http://www-personal.umich.edu/~mejn/netdata/}).
The variables are as follows:
}
\keyword{datasets}

