% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-net.R
\name{fortify.network}
\alias{fortify.network}
\title{Function for converting a network object into the correct format for use with geomnet}
\usage{
\method{fortify}{network}(model, data = NULL, group = NULL, ...)
}
\arguments{
\item{model}{object of class "network"}

\item{data}{NULL - not used in this function}

\item{group}{character. Used for facetting. If you wish to facet on an edge variable provide the name of that variable here.}

\item{...}{not used in this function}
}
\description{
Function for converting a network object into the correct format for use with geomnet
}
\examples{
# class network (sna, network, statnet packages)

library(network)
data(emon, package = "network")
fortify(emon$Cheyenne)

}

