% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-net.r, R/stat-net.r
\name{geom_net}
\alias{geom_net}
\alias{stat_net}
\title{Geom for network visualization}
\usage{
geom_net(mapping = NULL, data = NULL, stat = "net",
  position = "identity", show.legend = NA, inherit.aes = TRUE,
  alpha = 0.25, layout = "kamadakawai", layout.par = list(),
  vertices = NULL, vlabel = FALSE, vcolour = NULL, vsize = NULL,
  vshape = NULL, directed = FALSE, ...)

stat_net(mapping = NULL, data = NULL, geom = "point",
  position = "identity", show.legend = NA, inherit.aes = TRUE,
  layout = "kamadakawai", layout.par = list(), vertices = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}

\item{geom}{The geometric object to use display the data}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
   a warning. If \code{TRUE} silently removes missing values.}
}
\value{
A data frame with additional columns:
  \item{density}{density estimate}
  \item{scaled}{density estimate, scaled to maximum of 1}
  \item{count}{density * number of points - not sure about usefulness for vase plots}
  \item{vasewidth}{density scaled for the vase plot, according to area, counts
                     or to a constant maximum width}
  \item{n}{number of points}
  \item{width}{width of vase bounding box}
}
\description{
Geom for network visualization
}
\examples{
library(ggplot2)
data(blood)
p <- ggplot(data = blood$edges, aes(from_id = from, to_id = to))
p + geom_net()
p + geom_net(vertices = blood$vertices)
p + geom_net(vertices = blood$vertices, vlabel=blood$vertices$rho=="neg")
p + geom_net(vertices = blood$vertices, vcolour = "orange", layout = 'circle',
vlabel = TRUE, vsize = 6, directed = TRUE) +  theme_net

#Madmen Relationships
data(madmen)
p <- ggplot(data = madmen$edges, aes(from_id = Name1, to_id = Name2))
p + geom_net(vlabel= TRUE)
p + geom_net(vertices = madmen$vertices, vsize=5, vlabel= TRUE, colour="grey30",
             aes(vcolour=c("#FF69B4", "#0099ff")[as.numeric(madmen$vertices$Gender)]))
}

