% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme-net.r
\name{ggtheme}
\alias{ggtheme}
\alias{theme_net}
\title{Theme for network visualization}
\usage{
theme_net(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Themes set the general aspect of the plot such as the colour of the
background, gridlines, the size and colour of fonts.
theme_net provides access to the regular ggplot2 theme, but removes any
background, axes, and ensures an aspect ratio of 1 for better
viewing of networks and graphs.
}
\examples{
library(ggplot2)
data(blood)
bloodnet <- merge(blood$edges, blood$vertices, by.x="from", by.y="label", all=TRUE)
p <- ggplot(data = bloodnet, aes(from_id = from, to_id = to))
p + geom_net()
p + geom_net() + theme_net()
}

