% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lesmis}
\alias{lesmis}
\title{Coappearance network of characters in Les Miserables (undirected)}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of three variables of length 254:
\itemize{
  \item from: Character 1
  \item to: Character 2
  \item degree: number of times they appear together in a chapter of Les Miserables
}
\item the vertices data set consists of two variables with information on 77 characters:
\itemize{
  \item id: Character ID number
  \item label: Character name
}
}}
\usage{
lesmis
}
\description{
A list of two datasets, vertices and edges, containing data on characters and chapters in Victor Hugo's Les Miserables (see \url{http://www-personal.umich.edu/~mejn/netdata/}).
The variables are as follows:
}
\references{
D. E. Knuth, The Stanford GraphBase: A Platform for Combinatorial Computing, Addison-Wesley, Reading, MA (1993).
}
\keyword{datasets}

