% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{madmen}
\alias{madmen}
\title{Network of romantic relationships in the TV show Mad Men (undirected)}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of two variables of length 39:
\itemize{
  \item Name1, Name2: Factor variables containing names of characters with a sexual relationship
}
\item the vertices data set consists of three variables with information on 45 characters of the show:
\itemize{
  \item label: Factor variable with name of the character,
  \item Gender: Factor variable of the gender of the character.
}
}}
\usage{
madmen
}
\description{
A list of two datasets, vertices and edges, containing information on sexual relations in the TV show Mad Men.
This data set was first compiled by Winston Chang for the package gcookbook (under the same name) and was extended here to include
the gender of each of the characters.
The variables are as follows:
}
\references{
Winston Chang. 2013. R Graphics Cookbook. O'Reilly Media, Inc..
}
\keyword{datasets}
