% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{metro_map}
\alias{metro_map}
\title{Map of Washington DC Metro area}
\format{A data frame of three variables of length 1638400:
\itemize{
\item x: Longitudinal coordinates
\item y: Latitudinal coordinates
\item fill: fill color in HEX form for that tile location
}}
\usage{
metro_map
}
\description{
A dataset containing information to draw a map of Washington DC using \code{geom_tile}.
}
\keyword{datasets}

