% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{metro_map}
\alias{metro_map}
\title{Map of Washington DC Metro area}
\format{An object of class "ggmap" and "raster" containing a map of the Rockville, MD, area.}
\usage{
metro_map
}
\description{
A dataset containing information to draw a map of Rockville, MD, and vicinity using \code{ggmap}.
This information was pulled from Google Maps using the \pkg{ggmap} package.
}
\examples{
\dontrun{
library(ggmap)
data(metro_map)
ggmap(metro_map)
}
}
\references{
D. Kahle and H. Wickham. ggmap: Spatial Visualization with ggplot2.
The R Journal, 5(1), 144-161. \url{https://journal.r-project.org/archive/2013-1/kahle-wickham.pdf}
}
\keyword{datasets}
