% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{soccer}
\alias{soccer}
\title{High school boys' soccer games in Iowa (undirected)}
\format{A list of two data frames:
\itemize{
\item the edges data set consists of 11 variables of length 4484:
\itemize{
  \item sub: The subregion of the state the school is in (1-8)
  \item home: The home team school
  \item season: The year the game was played
  \item week: Week of the season the game was played
  \item date: Date game was played
  \item away: The away team school
  \item ha: Home team, away team, or neutral?
  \item result: Did the home team win, lose, or tie?
  \item score: Game final score, home team first
  \item diff: Home team score minus away team score
  \item same_div: An indicator variable that is 1 if the schools are in the same division and 0 otherwise
}
\item the vertices data set consists of two variables with information on 157 schools:
\itemize{
  \item div: Division school is in (1, 2, or 3A)
  \item label: School name
}
}}
\usage{
soccer
}
\description{
A list of two datasets, vertices and edges, containing data on boys' soccer games at Iowa high schools in the 2011-2014 seasons.
This dataset was compiled by Danny Bero (\email{bero.danny@gmail.com}).
The variables are as follows:
}
\examples{
# prep the data
soccernet <- merge(soccer$edges, soccer$vertices, by.x = "home",
                   by.y = "label", all = TRUE)

# create plot
ggplot(data = soccernet, aes(from_id = home, to_id = away)) +
  geom_net(aes(colour = div, group = div), ealpha = .25, layout = 'fruchtermanreingold') +
  facet_wrap(~season) +
  theme_net()
}
\keyword{datasets}

